+++ deep_translator/detection.py
@@ -7,6 +7,7 @@ from requests.exceptions import HTTPError
 # Module global config
 config = {"url": 'https://ws.detectlanguage.com/0.2/detect',"headers": {'User-Agent': 'Detect Language API Python Client 1.4.0','Authorization': 'Bearer {}',}}
 
+
 def get_request_body(text, api_key, *args, **kwargs):
     """
     send a request and return the response body parsed as dictionary