+++ tests/test_microsoft_trans.py
@@ -20,12 +20,12 @@ def test_microsoft_successful_post_mock(mock_request_post):
         r.json = json_func
         return r
     mock_request_post.return_value = res()
-    assert MicrosoftTranslator(api_key="an_api_key", target="en").translate("auf wiedersehen!") == "See you later!"
+    assert MicrosoftTranslator(api_key="an_api_key", source='de', target="en").translate("auf wiedersehen!") == "See you later!"
 
 
 def test_MicrosoftAPIerror():
     with pytest.raises(exceptions.MicrosoftAPIerror):
-        MicrosoftTranslator(api_key="empty", target="en").translate("text")
+        MicrosoftTranslator(api_key="empty", source='de', target="en").translate("text")
 
 
 # the remaining tests are actual requests to Microsoft API and use an api key
@@ -58,5 +58,5 @@ def test_incorrect_target_attributes():
 def test_abbreviations():
     m1 = MicrosoftTranslator(api_key=APIkey, source="en", target="fr")
     m2 = MicrosoftTranslator(api_key=APIkey, source="English", target="French")
-    assert ''.join(m1.source) == ''.join(m2.source)
-    assert ''.join(m1.target) == ''.join(m2.target)
+    assert ''.join(m1._source) == ''.join(m2._source)
+    assert ''.join(m1._target) == ''.join(m2._target)