+++ docs/history.rst
@@ -11,8 +11,8 @@ Important History
 - 1.3.4 bug fixes for the dutch language
 - 1.3.3 fixed bug in google translate
 
-- 1.2.5: added support for the DeepL translator
-- 1.2.4: added support for the QCRI translator
+- 1.2.5: added support for the DeeplTranslator translator
+- 1.2.4: added support for the QcriTranslator translator
 - 1.2.1: added support for the yandex translator
 - 1.1.9: fixed bug in requests
 - 1.1.5: added language detection